aider: Calculate the MD5 hash of the content and add it to the Metadata argument when uploading a file to S3.
# Aider chat conversation:

ASSISTANT: The `ContentMD5` argument is not supported by the `upload_fileobj()` function in the `boto3` library. However, you can manually calculate the MD5 hash of the content and include it in the `Metadata` argument when uploading the file to S3.

Here is the *SEARCH/REPLACE block* for the necessary change:

```python
klingon_file_manager/write.py
<<<<<<< SEARCH
                s3_client.upload_fileobj(
                    Fileobj=f,
                    Bucket=bucket_name,
                    Key=key,
                    ContentMD5=md5,
                    Callback=progress,
                    ExtraArgs={'Metadata': metadata}
                )
=======
                import hashlib
                md5_hash = hashlib.md5(content).hexdigest()
                metadata['ContentMD5'] = md5_hash
                s3_client.upload_fileobj(
                    Fileobj=f,
                    Bucket=bucket_name,
                    Key=key,
                    Callback=progress,
                    ExtraArgs={'Metadata': metadata}
                )
>>>>>>> REPLACE
```

This change calculates the MD5 hash of the content and adds it to the `Metadata` argument. The `ContentMD5` field in the metadata will now contain the MD5 hash of the content. Please note that this will not perform an MD5 hash check during the file upload. If you want to perform the MD5 hash check, you may need to implement it manually after the file upload.