aider: Modified `test.py` to set the `debug` argument to `True` when calling the `manage_file` function in `test_large_upload_progress`.
# Aider chat conversation:

USER: progress callback isn't showing when I add progress=True to `test_large_upload_progress` Do we need to change `test.py`?

ASSISTANT: Yes, we need to modify `test.py` to print the progress of the file upload. However, the `ProgressPercentage` class in `utils.py` is responsible for printing the progress. If it's not showing the progress, there might be an issue with this class.

Could you please add `klingon_file_manager/utils.py` to the chat so I can check it?

USER: I added these *read-write* files: klingon_file_manager/utils.py

ASSISTANT: Thank you for providing the `utils.py` file.

Upon reviewing the `ProgressPercentage` class, it seems to be correctly implemented. The progress should be printed to the console when the `__call__` method of the `ProgressPercentage` class is invoked during the file upload.

However, the `ProgressPercentage` class is only used in the `write_file` function in `write.py` when the `progress` argument is `True`.

Let's modify the `test.py` file to ensure that the `progress` argument is set to `True` when calling the `manage_file` function.

Here is the *SEARCH/REPLACE block* for the modification:

```python
test.py
<<<<<<< SEARCH
    result = manage_file(
        action='post',
        path="s3://fsg-gobbler/tests/"+file_name,
        content=file_content,
        md5=md5_hash_hex,
        metadata=metadata,
        progress=True,
        debug=False
    )
=======
    result = manage_file(
        action='post',
        path="s3://fsg-gobbler/tests/"+file_name,
        content=file_content,
        md5=md5_hash_hex,
        metadata=metadata,
        progress=True,
        debug=True
    )
>>>>>>> REPLACE
```

In this block, I've changed the `debug` argument from `False` to `True`. This will print more detailed information about the file upload process, which might help us understand why the progress isn't being shown.