aider: Call the progress callback with a dummy value for bytes_amount.
# Aider chat conversation:

USER: I ran this command:

 python test.py

And got this output:

100+0 records in
100+0 records out
104857600 bytes transferred in 0.013794 secs (7601681891 bytes/sec)
Traceback (most recent call last):
  File "/Users/djh/work/src/github.com/djh00t/module_klingon_file_manager/test.py", line 98, in <module>
    test_large_upload_progress()
  File "/Users/djh/work/src/github.com/djh00t/module_klingon_file_manager/test.py", line 80, in test_large_upload_progress
    progress_callback()
TypeError: __call__() missing 1 required positional argument: 'bytes_amount'

ASSISTANT: The error message indicates that the `__call__()` method of the `ProgressPercentage` class is missing a required positional argument: `bytes_amount`. This argument is supposed to represent the number of bytes that have been transferred so far.

To fix this issue, we need to modify the `test.py` file where the `ProgressPercentage` object is being called without the `bytes_amount` argument.

Here is the change that needs to be made:

```python
test.py
<<<<<<< SEARCH
    # Call the progress callback
    progress_callback()
=======
    # Call the progress callback with a dummy value for bytes_amount
    progress_callback(0)
>>>>>>> REPLACE
```

This change will call the `ProgressPercentage` object with a dummy value of `0` for `bytes_amount`. This is just a placeholder value, and the actual value of `bytes_amount` will be updated by the `boto3` library during the file upload process.